refactor: moving ResourceWithTable to types file
