+++ ui/src/lib/features/k8s/deployments/store.ts
@@ -3,8 +3,8 @@
 
 import type { V1Deployment as Resource } from '@kubernetes/client-node'
 
-import { ResourceStore, type ResourceStoreInterface, type ResourceWithTable } from '../store'
-import { type ColumnWrapper, type CommonRow } from '../types'
+import { ResourceStore, type ResourceStoreInterface } from '../store'
+import { type ColumnWrapper, type CommonRow, type ResourceWithTable } from '../types'
 
 interface Row extends CommonRow {
   ready: string
